Ankle PDOs working
Both ankles working. Code is a bit of a mess and requires editing. Currently only initialising the ankles and only driving a single ankle (movethroughtraj). Needs tidy up, but can actuate both ankles by only changing the single variable. Had problems with PDOs and reading from the OD,  but seems to have resovled itself. Was possibly due to me writing to the OD in one of my print statements.